+++ FirebaseSessions/Tests/Unit/Mocks/MockInstallationsProtocol.swift
@@ -17,7 +17,7 @@
 
 @testable import FirebaseSessions
 
-class MockInstallationsProtocol: InstallationsProtocol {
+class MockInstallationsProtocol: InstallationsProtocol, @unchecked Sendable {
   static let testInstallationId = "testInstallationId"
   static let testAuthToken = "testAuthToken"
   var result: Result<(String, String), Error> = .success((testInstallationId, testAuthToken))