+++ FirebaseVertexAI/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
-- [added] Added support for specifiying the minimum and maximum number of items
-  (`minItems` / `maxItems`) to generate in an array `Schema`.
+- [added] Added support for specifying the minimum and maximum number of items
+  (`minItems` / `maxItems`) to generate in an array `Schema`. (#14671)
 
 # 11.11.0
 - [added] Emits a warning when attempting to use an incompatible model with