+++ FirebaseVertexAI/CHANGELOG.md
@@ -6,8 +6,17 @@
   Note: This feature is in Public Preview and relies on experimental models,
   which means that it is not subject to any SLA or deprecation policy and could
   change in backwards-incompatible ways.
-- [added] Added support for specifying the minimum and maximum number of items
-  (`minItems` / `maxItems`) to generate in an array `Schema`. (#14671)
+- [added] Added support for multiple new `Schema` fields. (#14671)
+  - `minItems` / `maxItems`: Specifies the minimum and maximum number of items
+    to generate in an array `Schema`.
+  - `title`: A human-readable name/summary for the schema, helping to document
+    its purpose.
+  - `minimum` / `maximum`: Specifies the minimum or maximum values for numeric
+    types (`integer`, `float`, `double`).
+  - `anyOf`: Allows specifying that the generated data may be *any* of the
+    provided sub-schemas.
+  - `propertyOrdering`: Specifies the order for keys in the generated JSON
+    output for object schemas.
 - [fixed] Fixed an issue where network requests would fail in the iOS 18.4
   simulator due to a `URLSession` bug introduced in Xcode 16.3. (#14677)